+++ libs/telemetry/src/main/java/org/opensearch/telemetry/tracing/SpanScope.java
@@ -8,11 +8,15 @@
 
 package org.opensearch.telemetry.tracing;
 
+import org.opensearch.common.annotation.ExperimentalApi;
 import org.opensearch.telemetry.tracing.noop.NoopSpanScope;
 
 /**
  * An auto-closeable that represents scope of the span.
+ *
+ * @opensearch.experimental
  */
+@ExperimentalApi
 public interface SpanScope extends AutoCloseable {
 
     /**